chore(sdk/go): consolidate `redis` and `outboun_redis` into single package
This commit consolidates the two new `redis` and `outbound_redis` packages
from the Go SDK into a single `redis` package that contains both the Redis
trigger implementation and the functionality to connect to an external Redis
instance.
