--- sdk/go/outbound_redis/Makefile
@@ -1,7 +1,0 @@
-.PHONY: generate
-generate:
-	wit-bindgen c --import ../../../wit/ephemeral/outbound-redis.wit
-
-.PHONY: clean
-clean:
-	rm outbound-redis.h outbound-redis.c
