--- sdk/go/outbound_redis/README.md
@@ -1,3 +1,0 @@
-# Outbound Redis for Spin
-
-To generate the c bindings run `make generate`
