+++ sdk/go/redis/outbound-redis.c
